requirements.txt
Django==4.1.3
google-generativeai>=0.7.0
numpy>=1.23
